requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asttokens==3.0.0
backcall==0.2.0
bcrypt==4.2.1
cffi==1.17.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
dnspython==2.6.1
ecdsa==0.19.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
passlib==1.7.4
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pywin32==308
pyzmq==26.2.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
zipp==3.20.2
